+++ src/RcppExports.cpp
@@ -50,15 +50,15 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// colSumByFac
-arma::mat colSumByFac(SEXP sY, SEXP rowSel);
-RcppExport SEXP _sccore_colSumByFac(SEXP sYSEXP, SEXP rowSelSEXP) {
+// colSumByFactor
+NumericMatrix colSumByFactor(SEXP sY, IntegerVector rowSel);
+RcppExport SEXP _sccore_colSumByFactor(SEXP sYSEXP, SEXP rowSelSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< SEXP >::type sY(sYSEXP);
-    Rcpp::traits::input_parameter< SEXP >::type rowSel(rowSelSEXP);
-    rcpp_result_gen = Rcpp::wrap(colSumByFac(sY, rowSel));
+    Rcpp::traits::input_parameter< IntegerVector >::type rowSel(rowSelSEXP);
+    rcpp_result_gen = Rcpp::wrap(colSumByFactor(sY, rowSel));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -129,7 +129,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_sccore_as_factor", (DL_FUNC) &_sccore_as_factor, 1},
     {"_sccore_get_nearest_neighbors", (DL_FUNC) &_sccore_get_nearest_neighbors, 10},
     {"_sccore_jsDist", (DL_FUNC) &_sccore_jsDist, 2},
-    {"_sccore_colSumByFac", (DL_FUNC) &_sccore_colSumByFac, 2},
+    {"_sccore_colSumByFactor", (DL_FUNC) &_sccore_colSumByFactor, 2},
     {"_sccore_propagate_labels", (DL_FUNC) &_sccore_propagate_labels, 9},
     {"_sccore_smooth_count_matrix", (DL_FUNC) &_sccore_smooth_count_matrix, 10},
     {"_sccore_adjacentVertices", (DL_FUNC) &_sccore_adjacentVertices, 1},